+++ src/debug/objects.rs
@@ -83,7 +83,6 @@ fn draw_debug_gizmos(
         let Some(map_asset) = assets.get(&tiled_map.0) else {
             continue;
         };
-        // XXX: stil and issue on isometric maps for objects with a non-null rotation
         for (idx, (_, entity)) in storage.objects().enumerate() {
             let Ok((object, transform)) = object_query.get(*entity) else {
                 continue;