+++ CHANGELOG.md
@@ -10,9 +10,9 @@ A [migration guide](https://adrien-bon.github.io/bevy_ecs_tiled/migrations/v0_9.
 
 ### Changed
 
-- Rationalize usage of  `tile_size` vs. `grid_size`
-- Deduce the largest `tile_size` of a given map and use that to
-  perform map-level operations needing a `tile_size` instead of relying upong the `grid_size`
+- Rationalize usage of `tile_size` vs. `grid_size`
+- Deduce the largest `tile_size` of a given map and use that to perform map-level operations needing a `tile_size`
+  instead of relying upon the `grid_size`
 
 ### Bugfixes